third_party/blink/web_tests/fast/scrolling/scrollbar-mousedown-mouseup-expected.txt

This is a scrollable div.

PASS events.length is 4
PASS events[0].type is "mousedown"
PASS events[0].target.id is "scrollme"
PASS events[0].which is 1
PASS events[1].type is "mouseup"
PASS events[1].target.id is "scrollme"
PASS events[1].which is 1
PASS events[2].type is "mousedown"
PASS events[2].target.id is "scrollme"
PASS events[2].which is 2
PASS events[3].type is "mouseup"
PASS events[3].target.id is "scrollme"
PASS events[3].which is 2
PASS successfullyParsed is true

TEST COMPLETE
